+++ instrument_controls.py
